requirements.txt
streamlit>=1.55.0
pandas>=2.2.0
paho-mqtt>=2.1.0
pytest>=8.0.0
